+++ src/check/rules/unify.rs
@@ -1,5 +1,5 @@
-use crate::check::monad::{TCM, TCS};
-use crate::syntax::core::{Term, Val};
+use crate::check::monad::{TCE, TCM, TCS};
+use crate::syntax::core::{Elim, Term, Val};
 
 pub fn subtype(tcs: TCS, sub: &Val, sup: &Val) -> TCM {
     use Val::*;
@@ -9,11 +9,32 @@ pub fn subtype(tcs: TCS, sub: &Val, sup: &Val) -> TCM {
     }
 }
 
-pub fn unify(tcs: TCS, left: &Term, right: &Term) -> TCM {
+pub fn unify(mut tcs: TCS, left: &Term, right: &Term) -> TCM {
     use Term::*;
     match (left, right) {
         (Whnf(left), Whnf(right)) => unify_val(tcs, left, right),
-        _ => unimplemented!(),
+        (Redex(g0, args0), Redex(g1, args1)) if args0.len() == args1.len() => {
+            if g0 != g1 {
+                return Err(TCE::DifferentName(
+                    tcs.def(*g0).def_name().clone(),
+                    tcs.def(*g1).def_name().clone(),
+                ));
+            }
+            for (a, b) in args0.iter().zip(args1.iter()) {
+                tcs = unify_elim(tcs, a, b)?;
+            }
+            Ok(tcs)
+        }
+        (a, b) => Err(TCE::DifferentTerm(a.clone(), b.clone())),
+    }
+}
+
+pub fn unify_elim(tcs: TCS, left: &Elim, right: &Elim) -> TCM {
+    use Elim::*;
+    match (left, right) {
+        (Proj(a), Proj(b)) if a == b => Ok(tcs),
+        (App(a), App(b)) => unify(tcs, &**a, &**b),
+        (a, b) => Err(TCE::DifferentElim(a.clone(), b.clone())),
     }
 }